Add Rake tasks for sorbet and test suites
Now we have seperate test suites, we want seperate Rake tasks for
all of them. We also want one for Sorbet. Lastly, the default
command is changed to run everything in the same order as will be
used in Github Actions. It is a maven inspired 'fail fast'
order of

- Sorbet (static type-checking)
- Rubocop (static linting)
- Tests
  - Unit (with 90% coverage required)
  - Integration (with 70% coverage required)
  - Functional (with 50% coverage required)
- Documentation
  - Build
  - Verify coverage
  - Lint (this currently does not fail)