Add Tae::Command
Tae::Command is the base class which will be scanned for to add commands
to tae when it boots up. Right now the implmenentation has the ugly need
to forcefully look up instance variables by name as sorbet didn't
understand the approach used in dry-cli to create class methods for DSL
style class creation but still have nice accessors for the same
variables. Ease of use at the client side wins.